rename dagit-screenshot -> dagster-ui-screenshot
