+++ .buildkite/dagster-buildkite/dagster_buildkite/package_spec.py
@@ -33,7 +33,7 @@ _INFRASTRUCTURE_PACKAGES = [
     ".buildkite/dagster-buildkite",
     "python_modules/automation",
     "python_modules/dagster-test",
-    "python_modules/docs/dagit-screenshot",
+    "python_modules/docs/dagster-ui-screenshot",
 ]
