+++ .buildkite/dagster-buildkite/dagster_buildkite/steps/docs.py
@@ -5,7 +5,7 @@ from dagster_buildkite.steps.tox import build_tox_step
 from ..python_version import AvailablePythonVersion
 from ..step_builder import CommandStepBuilder
 from ..utils import BuildkiteLeafStep, BuildkiteStep, GroupStep, skip_if_no_docs_changes
-from .packages import build_dagit_screenshot_steps, build_example_packages_steps
+from .packages import build_dagster_ui_screenshot_steps, build_example_packages_steps
 
 
 def build_docs_steps() -> List[BuildkiteStep]:
@@ -59,6 +59,6 @@ def build_docs_steps() -> List[BuildkiteStep]:
     ]
 
     steps += build_example_packages_steps()
-    steps += build_dagit_screenshot_steps()
+    steps += build_dagster_ui_screenshot_steps()
 
     return steps
