+++ docs/dagster-ui-screenshot/dagster_ui_screenshot/cli.py
@@ -2,12 +2,16 @@ from typing import Optional
 
 import click
 
-from dagit_screenshot.commands.asset_svg import generate_svg as _svg
-from dagit_screenshot.commands.audit import audit as _audit
-from dagit_screenshot.commands.capture import capture as _capture
-from dagit_screenshot.commands.show import show as _show
-from dagit_screenshot.defaults import DEFAULT_OUTPUT_ROOT, DEFAULT_SPEC_DB, DEFAULT_WORKSPACE_ROOT
-from dagit_screenshot.utils import load_spec, normalize_output_path, spec_id_to_relative_path
+from dagster_ui_screenshot.commands.asset_svg import generate_svg as _svg
+from dagster_ui_screenshot.commands.audit import audit as _audit
+from dagster_ui_screenshot.commands.capture import capture as _capture
+from dagster_ui_screenshot.commands.show import show as _show
+from dagster_ui_screenshot.defaults import (
+    DEFAULT_OUTPUT_ROOT,
+    DEFAULT_SPEC_DB,
+    DEFAULT_WORKSPACE_ROOT,
+)
+from dagster_ui_screenshot.utils import load_spec, normalize_output_path, spec_id_to_relative_path
 
 
 @click.group(
@@ -33,13 +37,13 @@ from dagit_screenshot.utils import load_spec, normalize_output_path, spec_id_to_
     help="Path to root directory against which relative spec workspaces are resolved.",
 )
 @click.pass_context
-def dagit_screenshot(ctx, output_root: str, spec_db: str, workspace_root: str) -> None:
+def dagster_ui_screenshot(ctx, output_root: str, spec_db: str, workspace_root: str) -> None:
     ctx.obj["output_root"] = output_root
     ctx.obj["spec_db"] = spec_db
     ctx.obj["workspace_root"] = workspace_root
 
 
-@dagit_screenshot.command(help="""
+@dagster_ui_screenshot.command(help="""
     Audit a screenshot spec database. Verifies that screenshot specs are valid and that referenced
     workspace files exist. Optionally verify that corresponding output files exist.
     """)
@@ -59,7 +63,7 @@ def audit(ctx, verify_outputs) -> None:
     _audit(spec_db, output_root, workspace_root, verify_outputs=verify_outputs)
 
 
-@dagit_screenshot.command(help="Reads a screenshot spec and captures a screenshot")
+@dagster_ui_screenshot.command(help="Reads a screenshot spec and captures a screenshot")
 @click.argument("spec_id", nargs=1)
 @click.option(
     "--output-path",
@@ -79,7 +83,7 @@ def capture(ctx, spec_id: str, output_path: str) -> None:
     _capture(spec, output_path)
 
 
-@dagit_screenshot.command(help="Dump the contents of a screenshot DB to the terminal as YAML.")
+@dagster_ui_screenshot.command(help="Dump the contents of a screenshot DB to the terminal as YAML.")
 @click.option(
     "--prefix",
     help="If provided, only specs with ids starting with the passed value will be dumped.",
@@ -89,7 +93,7 @@ def show(ctx, prefix: Optional[str]):
     _show(ctx.obj["spec_db"], prefix)
 
 
-@dagit_screenshot.command(
+@dagster_ui_screenshot.command(
     help="Reads a given markdown file and generates asset graph SVGs for code snippets."
 )
 @click.option("--target", "-t", help="Path to markdown file to process.")
@@ -98,4 +102,4 @@ def svg(target: str) -> None:
 
 
 def main():
-    dagit_screenshot(obj={})
+    dagster_ui_screenshot(obj={})
