+++ docs/dagster-ui-screenshot/dagster_ui_screenshot/defaults.py
