+++ docs/tox.ini
@@ -39,6 +39,6 @@ commands =
 
 [testenv:audit-screenshots]
 deps =
-  -e ./dagit-screenshot
+  -e ./dagster-ui-screenshot
 commands =
-  dagit-screenshot audit --verify-outputs
+  dagster-ui-screenshot audit --verify-outputs
