+++ pyright/master/include.txt
@@ -1,4 +1,5 @@
 .buildkite/dagster-buildkite
+docs/dagster-ui-screenshot
 docs/sphinx/_ext/dagster-sphinx
 python_modules
 examples
