+++ pyright/master/requirements-pinned.txt
@@ -146,6 +146,7 @@ cycler==0.11.0
 -e python_modules/libraries/dagster-ssh
 -e python_modules/dagster-test
 -e python_modules/libraries/dagster-twilio
+-e docs/dagster-ui-screenshot
 -e python_modules/libraries/dagster-wandb
 -e python_modules/dagster-webserver
 -e python_modules/libraries/dagstermill
@@ -344,6 +345,7 @@ opentelemetry-semantic-conventions==0.36b0
 ordered-set==4.1.0
 orjson==3.9.5
 oscrypto==1.3.0
+outcome==1.2.0
 overrides==7.4.0
 packaging==23.1
 pandas==2.0.3
@@ -398,6 +400,7 @@ pyOpenSSL==23.2.0
 pyparsing==3.0.9
 pypd==1.1.0
 pyright==1.1.316
+PySocks==1.7.1
 pyspark==3.4.1
 pytablereader==0.31.4
 pytest==7.0.1
@@ -444,6 +447,7 @@ scikit-learn==1.3.0
 scipy==1.11.2
 scrapbook==0.5.0
 seaborn==0.12.2
+selenium==4.10.0
 Send2Trash==1.8.2
 sending==0.3.0
 sentry-sdk==1.29.2
@@ -503,6 +507,8 @@ tornado==6.3.3
 tox==3.25.0
 tqdm==4.66.1
 traitlets==5.9.0
+trio==0.22.2
+trio-websocket==0.10.3
 -e examples/tutorial_notebook_assets
 twilio==8.5.0
 twine==1.15.0
@@ -559,6 +565,7 @@ widgetsnbextension==4.0.8
 -e examples/with_wandb
 wordcloud==1.9.2
 wrapt==1.15.0
+wsproto==1.2.0
 WTForms==3.0.1
 xgboost==1.7.6
 xmltodict==0.12.0
