+++ pyright/master/requirements.txt
@@ -26,6 +26,7 @@
 
 ### LIBRARIES
 
+-e docs/dagster-ui-screenshot
 -e docs/sphinx/_ext/dagster-sphinx
 -e python_modules/automation
 -e python_modules/dagster[pyright,test]
